refactor: split SettingsPage into 7 tab components (650→98 lines)
Extract GeneralTab, ConnectivityTab, ModelsTab, PromptsTab, AccountTab,
MaintenanceTab, and AdvancedTab from the monolithic SettingsPage.vue.
Move settings-card to global stylesheet. Each tab owns its own logic and
confirmation modals where needed.
