+++ src/views/settings/AccountTab.vue
@@ -1,0 +1,9 @@
+<template>
+  <div class="flex flex-col gap-[8px]">
+    <AccountSettings />
+  </div>
+</template>
+
+<script setup lang="ts">
+import AccountSettings from "../../components/settings/AccountSettings.vue";
+</script>
